refactor(DEV-10629): move `<ScopedStyleProvider/>` to separate file
